+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapter.ts
@@ -9,7 +9,7 @@ import type {JestEnvironment} from '@jest/environment';
 import type {TestResult} from '@jest/test-result';
 import type {Config} from '@jest/types';
 import type {TestFileEvent} from 'jest-runner';
-import type {RuntimeType as Runtime} from 'jest-runtime';
+import type Runtime from 'jest-runtime';
 import type {SnapshotStateType} from 'jest-snapshot';
 import {deepCyclicCopy} from 'jest-util';